# core/generated/swagger-ui/schema-registry-api-spec.yaml
openapi: 3.0.1
info:
  title: Confluent Schema Registry
  version: v1
paths:
  /compatibility/subjects/{subject}/versions/{version}:
    post:
      summary: Test schema compatibility against a particular schema subject-version
      description: "Test input schema against a particular version of a subject's\
        \ schema for compatibility. The compatibility level applied for the check\
        \ is the configured compatibility level for the subject (http:get:: /config/(string:\
        \ subject)). If this subject's compatibility level was never changed, then\
        \ the global compatibility level applies (http:get:: /config)."
      operationId: testCompatibilityBySubjectName
      parameters:
      - name: subject
        in: path
        description: Subject of the schema version against which compatibility is
          to be tested
        required: true
        schema:
          type: string
      - name: version
        in: path
        description: "Version of the subject's schema against which compatibility\
          \ is to be tested. Valid values for versionId are between [1,2^31-1] or\
          \ the string \"latest\".\"latest\" checks compatibility of the input schema\
          \ with the last registered schema under the specified subject"
        required: true
        schema:
          type: string
      - name: verbose
        in: query
        description: Whether to return detailed error messages
        schema:
          type: boolean
      requestBody:
        description: Schema
        content:
          application/vnd.schemaregistry.v1+json:
            schema:
              $ref: '#/components/schemas/RegisterSchemaRequest'
          application/vnd.schemaregistry+json:
            schema:
              $ref: '#/components/schemas/RegisterSchemaRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterSchemaRequest'
          application/octet-stream:
            schema:
              $ref: '#/components/schemas/RegisterSchemaRequest'
        required: true
      responses:
        "200":
          description: Compatibility check result
          content:
            application/vnd.schemaregistry.v1+json:
              schema:
                $ref: '#/components/schemas/CompatibilityCheckResponse'
            application/vnd.schemaregistry+json; qs=0.9:
              schema:
                $ref: '#/components/schemas/CompatibilityCheckResponse'
            application/json; qs=0.5:
              schema:
                $ref: '#/components/schemas/CompatibilityCheckResponse'
        "404":
          description: |-
            Error code 40401 -- Subject not found
            Error code 40402 -- Version not found
        "422":
          description: |-
            Error code 42201 -- Invalid schema or schema type
            Error code 42202 -- Invalid version
        "500":
          description: Error code 50001 -- Error in the backend data store
...
  /compatibility/subjects/{subject}/versions:
    post:
      summary: Test schema compatibility against all schemas under a subject
      description: "Test input schema against a subject's schemas for compatibility,\
        \ based on the configured compatibility level of the subject. In other words,\
        \ it will perform the same compatibility check as register for that subject.\
        \ The compatibility level applied for the check is the configured compatibility\
        \ level for the subject (http:get:: /config/(string: subject)). If this subject's\
        \ compatibility level was never changed, then the global compatibility level\
        \ applies (http:get:: /config)."
      operationId: testCompatibilityForSubject
      parameters:
      - name: subject
        in: path
        description: Subject of the schema version against which compatibility is
          to be tested
        required: true
        schema:
          type: string
      - name: verbose
        in: query
        description: Whether to return detailed error messages
        schema:
          type: boolean
      requestBody:
        description: Schema
        content:
          application/vnd.schemaregistry.v1+json:
            schema:
              $ref: '#/components/schemas/RegisterSchemaRequest'
          application/vnd.schemaregistry+json:
            schema:
              $ref: '#/components/schemas/RegisterSchemaRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterSchemaRequest'
          application/octet-stream:
            schema:
              $ref: '#/components/schemas/RegisterSchemaRequest'
        required: true
      responses:
        "200":
          description: Compatibility check result
          content:
            application/vnd.schemaregistry.v1+json:
              schema:
                $ref: '#/components/schemas/CompatibilityCheckResponse'
            application/vnd.schemaregistry+json; qs=0.9:
              schema:
                $ref: '#/components/schemas/CompatibilityCheckResponse'
            application/json; qs=0.5:
              schema:
                $ref: '#/components/schemas/CompatibilityCheckResponse'
        "422":
          description: |-
            Error code 42201 -- Invalid schema or schema type
            Error code 42202 -- Invalid version
        "500":
          description: Error code 50001 -- Error in the backend data store
...
  /subjects/{subject}/versions:
    get:
      summary: List versions under subject
      description: Retrieves a list of versions registered under the specified subject.
      operationId: listVersions
      parameters:
      - name: subject
        in: path
        description: Name of the subject
        required: true
        schema:
          type: string
      - name: deleted
        in: query
        description: Whether to include deleted schemas
        schema:
          type: boolean
      - name: deletedOnly
        in: query
        description: Whether to return deleted schemas only
        schema:
          type: boolean
      responses:
        "200":
          description: The version numbers matching the specified parameters
          content:
            application/vnd.schemaregistry.v1+json:
              schema:
                type: array
                items:
                  type: integer
                  format: int32
            application/vnd.schemaregistry+json; qs=0.9:
              schema:
                type: array
                items:
                  type: integer
                  format: int32
            application/json; qs=0.5:
              schema:
                type: array
                items:
                  type: integer
                  format: int32
        "404":
          description: Error code 40401 -- Subject not found
        "500":
          description: Error code 50001 -- Error in the backend data store
    post:
      summary: Register schema under a subject
      description: |-
        Register a new schema under the specified subject. If successfully registered, this returns the unique identifier of this schema in the registry. The returned identifier should be used to retrieve this schema from the schemas resource and is different from the schema's version which is associated with the subject. If the same schema is registered under a different subject, the same identifier will be returned. However, the version of the schema may be different under different subjects.
        A schema should be compatible with the previously registered schema or schemas (if there are any) as per the configured compatibility level. The configured compatibility level can be obtained by issuing a GET http:get:: /config/(string: subject). If that returns null, then GET http:get:: /config
        When there are multiple instances of Schema Registry running in the same cluster, the schema registration request will be forwarded to one of the instances designated as the primary. If the primary is not available, the client will get an error code indicating that the forwarding has failed.
      operationId: register
      parameters:
      - name: subject
        in: path
        description: Name of the subject
        required: true
        schema:
          type: string
      - name: normalize
        in: query
        description: Whether to register the normalized schema
        schema:
          type: boolean
      requestBody:
        description: Schema
        content:
          application/vnd.schemaregistry.v1+json:
            schema:
              $ref: '#/components/schemas/RegisterSchemaRequest'
          application/vnd.schemaregistry+json:
            schema:
              $ref: '#/components/schemas/RegisterSchemaRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterSchemaRequest'
          application/octet-stream:
            schema:
              $ref: '#/components/schemas/RegisterSchemaRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/vnd.schemaregistry.v1+json:
              schema:
                $ref: '#/components/schemas/RegisterSchemaResponse'
            application/vnd.schemaregistry+json; qs=0.9:
              schema:
                $ref: '#/components/schemas/RegisterSchemaResponse'
            application/json; qs=0.5:
              schema:
                $ref: '#/components/schemas/RegisterSchemaResponse'
        "409":
          description: Incompatible schema
        "422":
          description: Error code 42201 -- Invalid schema or schema type
        "500":
          description: |-
            Error code 50001 -- Error in the backend data store
            Error code 50002 -- Operation timed out
            Error code 50003 -- Error while forwarding the request to the primary
  /subjects/{subject}/versions/{version}:
    get:
      summary: Get schema by version
      description: Retrieves a specific version of the schema registered under this
        subject.
      operationId: getSchemaByVersion
      parameters:
      - name: subject
        in: path
        description: Name of the subject
        required: true
        schema:
          type: string
      - name: version
        in: path
        description: "Version of the schema to be returned. Valid values for versionId\
          \ are between [1,2^31-1] or the string \"latest\". \"latest\" returns the\
          \ last registered schema under the specified subject. Note that there may\
          \ be a new latest schema that gets registered right after this request is\
          \ served."
        required: true
        schema:
          type: string
      - name: deleted
        in: query
        description: Whether to include deleted schema
        schema:
          type: boolean
      responses:
        "200":
          description: The schema
          content:
            application/vnd.schemaregistry.v1+json:
              schema:
                $ref: '#/components/schemas/Schema'
            application/vnd.schemaregistry+json; qs=0.9:
              schema:
                $ref: '#/components/schemas/Schema'
            application/json; qs=0.5:
              schema:
                $ref: '#/components/schemas/Schema'
        "404":
          description: |-
            Error code 40401 -- Subject not found
            Error code 40402 -- Version not found
          content:
            application/vnd.schemaregistry.v1+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
            application/vnd.schemaregistry+json; qs=0.9:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
            application/json; qs=0.5:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "422":
          description: Error code 42202 -- Invalid version
          content:
            application/vnd.schemaregistry.v1+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
            application/vnd.schemaregistry+json; qs=0.9:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
            application/json; qs=0.5:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "500":
          description: Error code 50001 -- Error in the backend data store
          content:
            application/vnd.schemaregistry.v1+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
            application/vnd.schemaregistry+json; qs=0.9:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
            application/json; qs=0.5:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
    delete:
      summary: Delete schema version
      description: "Deletes a specific version of the schema registered under this\
        \ subject. This only deletes the version and the schema ID remains intact\
        \ making it still possible to decode data using the schema ID. This API is\
        \ recommended to be used only in development environments or under extreme\
        \ circumstances where-in, its required to delete a previously registered schema\
        \ for compatibility purposes or re-register previously registered schema."
      operationId: deleteSchemaVersion
      parameters:
      - name: subject
        in: path
        description: Name of the subject
        required: true
        schema:
          type: string
      - name: version
        in: path
        description: "Version of the schema to be returned. Valid values for versionId\
          \ are between [1,2^31-1] or the string \"latest\". \"latest\" returns the\
          \ last registered schema under the specified subject. Note that there may\
          \ be a new latest schema that gets registered right after this request is\
          \ served."
        required: true
        schema:
          type: string
      - name: permanent
        in: query
        description: Whether to perform a permanent delete
        schema:
          type: boolean
      responses:
        "200":
          description: Operation succeeded. Returns the schema version
          content:
            application/vnd.schemaregistry.v1+json:
              schema:
                type: integer
                format: int32
            application/vnd.schemaregistry+json; qs=0.9:
              schema:
                type: integer
                format: int32
            application/json; qs=0.5:
              schema:
                type: integer
                format: int32
        "404":
          description: |-
            Error code 40401 -- Subject not found
            Error code 40402 -- Version not found
        "422":
          description: Error code 42202 -- Invalid version
        "500":
          description: Error code 50001 -- Error in the backend data store
...
components:
  schemas:
    CompatibilityCheckResponse:
      type: object
      properties:
        is_compatible:
          type: boolean
          description: Whether the compared schemas are compatible
        messages:
          type: array
          description: Error messages
          items:
            type: string
            description: Error messages
      description: Compatibility check response
    RegisterSchemaRequest:
      type: object
      properties:
        version:
          type: integer
          description: Version number
          format: int32
        id:
          type: integer
          description: Globally unique identifier of the schema
          format: int32
        schemaType:
          type: string
          description: Schema type
        references:
          type: array
          description: References to other schemas
          items:
            $ref: '#/components/schemas/SchemaReference'
        schema:
          type: string
          description: Schema definition string
      description: Schema register request
    SchemaReference:
      type: object
      properties:
        name:
          type: string
          description: Reference name
        subject:
          type: string
          description: Name of the referenced subject
        version:
          type: integer
          description: Version number of the referenced subject
          format: int32
      description: Schema reference
...
    Schema:
      type: object
      properties:
        subject:
          type: string
          description: Name of the subject
        version:
          type: integer
          description: Version number
          format: int32
        id:
          type: integer
          description: Globally unique identifier of the schema
          format: int32
        schemaType:
          type: string
          description: Schema type
        references:
          type: array
          description: References to other schemas
          items:
            $ref: '#/components/schemas/SchemaReference'
        schema:
          type: string
          description: Schema definition string
      description: Schema
...
    RegisterSchemaResponse:
      type: object
      properties:
        id:
          type: integer
          description: Globally unique identifier of the schema
          format: int32
      description: Schema register response
    ErrorMessage:
      type: object
      properties:
        error_code:
          type: integer
          format: int32
        message:
          type: string
...